Decompile unk_02004A44 (sound system utilities, 95 functions)
Converts the game's core sound system module (2672 asm lines, 95 functions)
to matching C. Includes BGM/SE scene management, GB Sounds mapping,
microphone control, wave output, and sound state save/restore.

+++ lib/include/nnsys/snd/player.h
@@ -13,6 +13,6 @@ void NNS_SndPlayerSetPlayableSeqCount(int playerNo, int seqCount);
 BOOL NNS_SndPlayerCreateHeap(int playerNo, NNSSndHeapHandle heap, u32 size);
 void NNS_SndPlayerPauseByPlayerNo(int playerNo, BOOL flag);
 int NNS_SndPlayerCountPlayingSeqByPlayerNo(int playerNo);
-void NNS_SndPlayerSetPlayerVolume(int playerNo, u8 volume);
+void NNS_SndPlayerSetPlayerVolume(int playerNo, int volume);
 
 #endif // GUARD_PLAYER_H
